Integration of website material in hft.org
The website was being produced outside the project repo (for most github projects, the website is contained in the gh-pages branch).  I moved the website material, which consists mostly of duplicated material from hft.org and parameters and snippets used in the website production process (o-blog which is an org file to website technology).  This material is now in the gh-pages production section where you can also find some instructions about how to add blog material (it's easy!).

The new assets directory is associated with website production, and includes all the charts and miscellaneous bric-a-brac you need for a website.

Some more refactoring is possible as there is now duplicated material in the hft.org file, but I'll leave that for another day.