Added tangle/de-tangle hook in org file.
For all code files there is now a comment line on the first line that is how org-mode remembers where the code came from.

This now means that edits/pull requests can be made on the code files and this will be able to back-propagate to the org file code blocks. Edits can also be made to the org-file directly and an org-tangle will push these edits to the code files.

Which of these spots has definitive priority becomes somewhat murky as a result. If the org-mode file was declared as the ultimate reference, then we could simply delete the code files and expect people to run emacs and tangle the files anew each time they pull the repo.  Yuck.  If it causes ambiguity issues down the track, the solution is probably to make the code files dominant, and have the org blocks as a detangle only, one-way street.

For the good news, if you are browsing through the code files, and you are running emacs, you can jump back and forth from org to the code files quite easily.